Move only usage of `take_static_root_alloc` to its definition and inline it
